Add bulk tests

+++ Domain/Models/Electricity.cs
@@ -2,11 +2,11 @@
 {
     public class Electricity
     {
-        public long? Id { get; set; }
-        public long? RegionId { get; set; }
+        public int? Id { get; set; }
+        public int? RegionId { get; set; }
         public string Pavadinimas { get; set; } = null!;
         public string Tipas { get; set; } = null!;
-        public long? Numeris { get; set; }
+        public int? Numeris { get; set; }
         public decimal? Ppliusas { get; set; }
         public DateTime? Plt { get; set; }
         public decimal? Pminusas { get; set; }
