+++ Domain/Models/Region.cs
@@ -2,7 +2,7 @@
 {
     public class Region
     {
-        public long RegionId { get; set; }
+        public int RegionId { get; set; }
         public string RegionName { get; set; }
     }
 }
